use 'size_t' instead of 'unsigned long'
64비트 운영체제에서 unsigned long과 unsigned long long은 8byte로 같아 포인터 값을 전부 표현할 수 있지만, 다른 환경도 고려해야 하므로 size_t로 변경

+++ ft_printf/ft_print_pointer.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   ft_print_pointer.c                                 :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: seohyuki <seohyuki@student.42seoul.kr>     +#+  +:+       +#+        */
+/*   By: seohyeon <seohyeon@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/10 22:15:53 by seohyuki          #+#    #+#             */
-/*   Updated: 2022/09/10 22:35:40 by seohyuki         ###   ########.fr       */
+/*   Updated: 2022/09/12 19:59:27 by seohyeon         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-int	ft_putnbr_pointer_fd(unsigned long n)
+int	ft_putnbr_pointer_fd(size_t n)
 {
 	int	length;
 
@@ -27,7 +27,7 @@ int	ft_putnbr_pointer_fd(unsigned long n)
 	return (length);
 }
 
-int	ft_print_pointer(unsigned long ptr)
+int	ft_print_pointer(size_t ptr)
 {
 	int	length;
